# docs/openapi/sbom.openapi.yaml
openapi: 3.0.3
info:
  title: Vecta KMS SBOM and CBOM Service API
  version: 1.0.0
  description: OpenAPI contract for SBOM generation, vulnerability correlation,
    offline advisories, and CBOM/PQC readiness. Served at `/svc/sbom` through
    the Envoy edge.
servers:
  - url: /svc/sbom
    description: Envoy edge
...
paths:
  /sbom/generate:
    post:
      tags:
        - SBOM
      operationId: generateSBOM
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateSBOMRequest"
      responses:
        "202":
          description: SBOM generation accepted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - snapshot
                  - request_id
                properties:
                  status:
                    type: string
                    enum:
                      - accepted
                  snapshot:
                    $ref: "#/components/schemas/SBOMSnapshot"
                  request_id:
                    type: string
        "401":
          description: No valid bearer token (reason `unauthenticated`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Missing permission, or a tenant other than the token's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Unhandled SBOM generation failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      description: "Permission `sbom.write`. Audited as
        `audit.sbom.sbom_generate_requested`, refusals included (`reason` =
        `unauthenticated`, `permission_denied`, `tenant_mismatch`,
        `tenant_conflict`). The platform SBOM is shared by every tenant: only
        the platform tenant (or a tenant-less root token or kms-* service
        principal) may change it; anyone else is refused with reason
        `platform_tenant_required`."
  /sbom/latest:
    get:
      tags:
        - SBOM
      operationId: getLatestSBOM
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
      responses:
        "200":
          description: Latest SBOM snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SBOMSnapshotEnvelope"
        "401":
          description: No valid bearer token (reason `unauthenticated`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Missing permission, or a tenant other than the token's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: No SBOM snapshot found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Unhandled SBOM retrieval failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      description: Permission `sbom.read`. Audited as `audit.sbom.sbom_latest_read`,
        refusals included (`reason` = `unauthenticated`, `permission_denied`,
        `tenant_mismatch`, `tenant_conflict`).
  /sbom/history:
    get:
      tags:
        - SBOM
      operationId: listSBOMHistory
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Historical SBOM snapshots.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SBOMSnapshotListEnvelope"
        "401":
          description: No valid bearer token (reason `unauthenticated`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Missing permission, or a tenant other than the token's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Unhandled SBOM history failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      description: Permission `sbom.read`. Audited as
        `audit.sbom.sbom_history_listed`, refusals included (`reason` =
        `unauthenticated`, `permission_denied`, `tenant_mismatch`,
        `tenant_conflict`).
  /sbom/vulnerabilities:
    get:
      tags:
        - SBOM
      operationId: listSBOMVulnerabilities
      responses:
        "200":
          description: Merged Manual OSV, OSV, and Trivy findings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VulnerabilityListEnvelope"
        "401":
          description: No valid bearer token (reason `unauthenticated`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Missing permission, or a tenant other than the token's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: No SBOM snapshot found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Unhandled vulnerability correlation failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      description: Permission `sbom.read`. Audited as
        `audit.sbom.sbom_vulnerabilities_listed`, refusals included (`reason` =
        `unauthenticated`, `permission_denied`, `tenant_mismatch`,
        `tenant_conflict`).
  /sbom/advisories:
    get:
      tags:
        - SBOM Advisories
      operationId: listManualAdvisories
      responses:
        "200":
          description: Manual offline advisories.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualAdvisoryListEnvelope"
        "401":
          description: No valid bearer token (reason `unauthenticated`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Missing permission, or a tenant other than the token's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Unhandled advisory retrieval failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      description: Permission `sbom.read`. Audited as
        `audit.sbom.sbom_advisories_listed`, refusals included (`reason` =
        `unauthenticated`, `permission_denied`, `tenant_mismatch`,
        `tenant_conflict`).
    post:
      tags:
        - SBOM Advisories
      operationId: saveManualAdvisory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualAdvisoryUpsert"
      responses:
        "202":
          description: Manual advisory stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualAdvisoryEnvelope"
        "400":
          description: Advisory validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: No valid bearer token (reason `unauthenticated`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Missing permission, or a tenant other than the token's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Unhandled advisory persistence failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      description: "Permission `sbom.write`. Audited as
        `audit.sbom.sbom_advisory_saved`, refusals included (`reason` =
        `unauthenticated`, `permission_denied`, `tenant_mismatch`,
        `tenant_conflict`). The platform SBOM is shared by every tenant: only
        the platform tenant (or a tenant-less root token or kms-* service
        principal) may change it; anyone else is refused with reason
        `platform_tenant_required`."
  /sbom/advisories/{id}:
    delete:
      tags:
        - SBOM Advisories
      operationId: deleteManualAdvisory
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Manual advisory deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteEnvelope"
        "401":
          description: No valid bearer token (reason `unauthenticated`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Missing permission, or a tenant other than the token's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Manual advisory not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Unhandled advisory deletion failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      description: "Permission `sbom.delete`. Audited as
        `audit.sbom.sbom_advisory_deleted`, refusals included (`reason` =
        `unauthenticated`, `permission_denied`, `tenant_mismatch`,
        `tenant_conflict`). The platform SBOM is shared by every tenant: only
        the platform tenant (or a tenant-less root token or kms-* service
        principal) may change it; anyone else is refused with reason
        `platform_tenant_required`."
  /sbom/diff:
    get:
      tags:
        - SBOM
      operationId: diffSBOM
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/SnapshotFrom"
        - $ref: "#/components/parameters/SnapshotTo"
      responses:
        "200":
          description: SBOM diff result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiffEnvelope"
        "400":
          description: Missing snapshot ids.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: No valid bearer token (reason `unauthenticated`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Missing permission, or a tenant other than the token's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: One or both SBOM snapshots were not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Unhandled SBOM diff failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      description: Permission `sbom.read`. Audited as `audit.sbom.sbom_diff_read`,
        refusals included (`reason` = `unauthenticated`, `permission_denied`,
        `tenant_mismatch`, `tenant_conflict`).
  /sbom/{id}/export:
    get:
      tags:
        - SBOM
      operationId: exportSBOM
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/SBOMExportFormat"
        - $ref: "#/components/parameters/SBOMExportEncoding"
      responses:
        "200":
          description: SBOM export payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportEnvelope"
        "400":
          description: Unsupported export format.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: No valid bearer token (reason `unauthenticated`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Missing permission, or a tenant other than the token's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: SBOM snapshot not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Unhandled SBOM export failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      description: Permission `sbom.read`. Audited as `audit.sbom.sbom_exported`,
        refusals included (`reason` = `unauthenticated`, `permission_denied`,
        `tenant_mismatch`, `tenant_conflict`).
  /sbom/{id}:
    get:
      tags:
        - SBOM
      operationId: getSBOMById
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: SBOM snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SBOMSnapshotEnvelope"
        "401":
          description: No valid bearer token (reason `unauthenticated`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Missing permission, or a tenant other than the token's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: SBOM snapshot not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Unhandled SBOM retrieval failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      description: Permission `sbom.read`. Audited as `audit.sbom.sbom_read`, refusals
        included (`reason` = `unauthenticated`, `permission_denied`,
        `tenant_mismatch`, `tenant_conflict`).
  /cbom/generate:
    post:
      tags:
        - CBOM
      operationId: generateCBOM
      parameters: &a1
        - &a2
          $ref: "#/components/parameters/RequestIdHeader"
        - &a3
          $ref: "#/components/parameters/TenantQuery"
        - &a4
          $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCBOMRequest"
      responses:
        "202":
          description: CBOM generation accepted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - snapshot
                  - request_id
                properties:
                  status:
                    type: string
                    enum:
                      - accepted
                  snapshot:
                    $ref: "#/components/schemas/CBOMSnapshot"
                  request_id:
                    type: string
        "400":
          description: Missing tenant scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: No valid bearer token (reason `unauthenticated`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Missing permission, or a tenant other than the token's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Unhandled CBOM generation failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      description: Permission `sbom.write`. Audited as
        `audit.sbom.cbom_generate_requested`, refusals included (`reason` =
        `unauthenticated`, `permission_denied`, `tenant_mismatch`,
        `tenant_conflict`). The CBOM is built for the token's tenant;
        `tenant_id` is no longer read from the body.
  /cbom/latest:
    get:
      tags:
        - CBOM
      operationId: getLatestCBOM
      parameters: *a1
      responses:
        "200":
          description: Latest CBOM snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CBOMSnapshotEnvelope"
        "400":
          description: Missing tenant scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: No valid bearer token (reason `unauthenticated`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Missing permission, or a tenant other than the token's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: No CBOM snapshot found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Unhandled CBOM retrieval failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      description: Permission `sbom.read`. Audited as `audit.sbom.cbom_latest_read`,
        refusals included (`reason` = `unauthenticated`, `permission_denied`,
        `tenant_mismatch`, `tenant_conflict`).
  /cbom/history:
    get:
      tags:
        - CBOM
      operationId: listCBOMHistory
      parameters:
        - *a2
        - *a3
        - *a4
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Historical CBOM snapshots.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CBOMSnapshotListEnvelope"
        "400":
          description: Missing tenant scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: No valid bearer token (reason `unauthenticated`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Missing permission, or a tenant other than the token's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Unhandled CBOM history failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      description: Permission `sbom.read`. Audited as
        `audit.sbom.cbom_history_listed`, refusals included (`reason` =
        `unauthenticated`, `permission_denied`, `tenant_mismatch`,
        `tenant_conflict`).
...
components:
  parameters:
    RequestIdHeader:
      name: X-Request-ID
      in: header
      required: false
      description: Optional client correlation id. The service generates one when omitted.
      schema:
        type: string
...
    LimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        example: 20
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - request_id
            - tenant_id
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            tenant_id:
              type: string
...
    SBOMSnapshot:
      type: object
      required:
        - id
        - source_hash
        - created_at
        - document
        - summary
      properties:
        id:
          type: string
        source_hash:
          type: string
        created_at: *a7
        document:
          $ref: "#/components/schemas/SBOMDocument"
        summary:
          $ref: "#/components/schemas/SBOMSummary"
...
    GenerateSBOMRequest:
      type: object
      properties:
        trigger:
          type: string
...
    SBOMSnapshotEnvelope:
      type: object
      required:
        - item
        - request_id
      properties:
        item:
          $ref: "#/components/schemas/SBOMSnapshot"
        request_id:
          type: string
    SBOMSnapshotListEnvelope:
      type: object
      required:
        - items
        - request_id
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SBOMSnapshot"
        request_id:
          type: string
    VulnerabilityListEnvelope:
      type: object
      required:
        - items
        - request_id
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VulnerabilityMatch"
        request_id:
          type: string
    ManualAdvisoryEnvelope:
      type: object
      required:
        - item
        - request_id
      properties:
        item:
          $ref: "#/components/schemas/ManualAdvisory"
        request_id:
          type: string
    ManualAdvisoryListEnvelope:
      type: object
      required:
        - items
        - request_id
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ManualAdvisory"
        request_id:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []